+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - :star: **New Feature:** New APIs to get and set Metadata on various objects in system
 - :star: **New Feature:** Beta support for publishing datasets to a Globus endpoint
 - :star: **New Feature:** Beta support for downloading published datasets using Globus
+- :star: **New Feature:** Owners of projects can now transfer their project ownership to another project user
 - :star: **Improvement:** Add notes as an ignored keyword for spreadsheet processing
 
 ## May 2019